refactor: extract shared @igniter/tx-verify pure package
decideVerification is imported by a Temporal workflow (value import) so it cannot
transit @igniter/pocket (cosmjs would break the sandbox). Move VerifyOutcome,
SupplierEffect, and decideVerification into a dependency-free @igniter/tx-verify
consumed by pocket and both workflow apps. pocket re-exports the types for
back-compat. Sandbox-safe for the verifier workflow.

(cherry picked from commit d8ca29f8409ea6eb5c39bc51d84a81e7be5f7930)
