+++ packages/tx-verify/eslint.config.mjs
@@ -1,0 +1,4 @@
+import { config } from "@igniter/eslint-config/base";
+
+/** @type {import("eslint").Linter.Config} */
+export default config;
